rename lib/asm.js and lib/dns.js
