+++ lib/asm.js
