+++ lib/asm/test.js
@@ -1,4 +1,4 @@
-import { Assembler, Registers } from 'lib/asm/asm.js'
+import { Assembler, Registers } from 'lib/asm.js'
 import { exec } from 'lib/proc.js'
 import { write_flags, write_mode } from 'lib/fs.js'
