+++ lib/dns.js
