+++ lib/ffi.js
@@ -1,4 +1,4 @@
-import { Assembler, Compiler, Registers } from 'lib/asm/asm.js'
+import { Assembler, Compiler, Registers } from 'lib/asm.js'
 
 const { ptr, core, wrap, assert } = lo
 const {
