+++ lib/tcc.js
@@ -1,4 +1,4 @@
-//import * as asm from 'lib/asm/asm.js'
+//import * as asm from 'lib/asm.js'
 //import { dump } from 'lib/binary.js'
 
 const { tcc } = lo.load('tcc')
